requirements.txt
-i https://pypi.org/simple
absl-py==2.1.0 ; python_version >= '3.7'
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==4.3.0 ; python_version >= '3.8'
argon2-cffi==23.1.0 ; python_version >= '3.7'
argon2-cffi-bindings==21.2.0 ; python_version >= '3.6'
arrow==1.3.0 ; python_version >= '3.8'
asttokens==2.4.1
async-lru==2.0.4 ; python_version >= '3.8'
attrs==23.2.0 ; python_version >= '3.7'
babel==2.14.0 ; python_version >= '3.7'
beautifulsoup4==4.12.3 ; python_full_version >= '3.6.0'
bleach==6.1.0 ; python_version >= '3.8'
boto3==1.34.59
botocore==1.34.59 ; python_version >= '3.8'
bottleneck==1.3.8
certifi==2024.2.2 ; python_version >= '3.6'
cffi==1.16.0 ; python_version >= '3.8'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
comm==0.2.1 ; python_version >= '3.8'
contourpy==1.2.0 ; python_version >= '3.9'
cycler==0.12.1 ; python_version >= '3.8'
debugpy==1.8.1 ; python_version >= '3.8'
decorator==5.1.1 ; python_version >= '3.5'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
executing==2.0.1 ; python_version >= '3.5'
fastai==1.0.61
fastapi==0.110.0
fastjsonschema==2.19.1
fastprogress==1.0.3 ; python_version >= '3.6'
filelock==3.13.1 ; python_version >= '3.8'
flatbuffers==24.3.7
fonttools==4.49.0 ; python_version >= '3.8'
fqdn==1.5.1
fsspec==2024.2.0 ; python_version >= '3.8'
h11==0.14.0 ; python_version >= '3.7'
httpcore==1.0.4 ; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0 ; python_version >= '3.8'
idna==3.6 ; python_version >= '3.5'
ipykernel==6.29.3 ; python_version >= '3.8'
ipython==8.22.2 ; python_version >= '3.10'
ipywidgets==8.1.2 ; python_version >= '3.7'
isoduration==20.11.0
jedi==0.19.1 ; python_version >= '3.6'
jinja2==3.1.3 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
json5==0.9.22 ; python_version >= '3.8'
jsonpointer==2.4
jsonschema==4.21.1 ; python_version >= '3.8'
jsonschema-specifications==2023.12.1 ; python_version >= '3.8'
jupyter==1.0.0
jupyter-client==8.6.0 ; python_version >= '3.8'
jupyter-console==6.6.3 ; python_version >= '3.7'
jupyter-core==5.7.1 ; python_version >= '3.8'
jupyter-events==0.9.0 ; python_version >= '3.8'
jupyter-lsp==2.2.4 ; python_version >= '3.8'
jupyter-server==2.13.0 ; python_version >= '3.8'
jupyter-server-terminals==0.5.2 ; python_version >= '3.8'
jupyterlab==4.1.4 ; python_version >= '3.8'
jupyterlab-pygments==0.3.0 ; python_version >= '3.8'
jupyterlab-server==2.25.3 ; python_version >= '3.8'
jupyterlab-widgets==3.0.10 ; python_version >= '3.7'
kiwisolver==1.4.5 ; python_version >= '3.7'
markupsafe==2.1.5 ; python_version >= '3.7'
matplotlib==3.8.3 ; python_version >= '3.9'
matplotlib-inline==0.1.6 ; python_version >= '3.5'
mediapipe==0.10.9
mistune==3.0.2 ; python_version >= '3.7'
mpmath==1.3.0
nbclient==0.9.0 ; python_full_version >= '3.8.0'
nbconvert==7.16.2 ; python_version >= '3.8'
nbformat==5.9.2 ; python_version >= '3.8'
nest-asyncio==1.6.0 ; python_version >= '3.5'
networkx==3.2.1 ; python_version >= '3.9'
notebook==7.1.1 ; python_version >= '3.8'
notebook-shim==0.2.4 ; python_version >= '3.7'
numexpr==2.9.0 ; python_version >= '3.9'
numpy==1.26.4 ; python_version >= '3.9'
nvidia-cublas-cu12==12.1.3.1 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-cupti-cu12==12.1.105 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu12==12.1.105 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-runtime-cu12==12.1.105 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cudnn-cu12==8.9.2.26 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cufft-cu12==11.0.2.54 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-curand-cu12==10.3.2.106 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusolver-cu12==11.4.5.107 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparse-cu12==12.1.0.106 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.19.3 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvjitlink-cu12==12.4.99 ; python_version >= '3'
nvidia-nvtx-cu12==12.1.105 ; platform_system == 'Linux' and platform_machine == 'x86_64'
opencv-contrib-python==4.9.0.80 ; python_version >= '3.6'
opencv-python==4.9.0.80
overrides==7.7.0 ; python_version >= '3.6'
packaging==23.2 ; python_version >= '3.7'
pandas==2.2.1 ; python_version >= '3.9'
pandocfilters==1.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3 ; python_version >= '3.6'
pexpect==4.9.0 ; sys_platform != 'win32' and sys_platform != 'emscripten'
pillow==10.2.0 ; python_version >= '3.8'
platformdirs==4.2.0 ; python_version >= '3.8'
prometheus-client==0.20.0 ; python_version >= '3.8'
prompt-toolkit==3.0.43 ; python_full_version >= '3.7.0'
protobuf==3.20.3 ; python_version >= '3.7'
psutil==5.9.8 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.6.3 ; python_version >= '3.8'
pydantic-core==2.16.3 ; python_version >= '3.8'
pygments==2.17.2 ; python_version >= '3.7'
pyparsing==3.1.2 ; python_full_version >= '3.6.8'
python-dateutil==2.9.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1
python-json-logger==2.0.7 ; python_version >= '3.6'
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1 ; python_version >= '3.6'
pyzmq==25.1.2 ; python_version >= '3.6'
qtconsole==5.5.1 ; python_version >= '3.8'
qtpy==2.4.1 ; python_version >= '3.7'
referencing==0.33.0 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
rfc3339-validator==0.1.4 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986-validator==0.1.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rpds-py==0.18.0 ; python_version >= '3.8'
s3transfer==0.10.0 ; python_version >= '3.8'
scipy==1.12.0 ; python_version >= '3.9'
seaborn==0.13.2 ; python_version >= '3.8'
send2trash==1.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
shapely==2.0.3 ; python_version >= '3.9'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1 ; python_version >= '3.7'
sounddevice==0.4.6 ; python_version >= '3.7'
soupsieve==2.5 ; python_version >= '3.8'
stack-data==0.6.3
starlette==0.36.3 ; python_version >= '3.8'
sympy==1.12 ; python_version >= '3.8'
terminado==0.18.0 ; python_version >= '3.8'
thop==0.1.1.post2209072238
tinycss2==1.2.1 ; python_version >= '3.7'
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4 ; python_version >= '3.8'
tqdm==4.66.2 ; python_version >= '3.7'
traitlets==5.14.1 ; python_version >= '3.8'
triton==2.2.0 ; python_version < '3.12' and platform_system == 'Linux' and platform_machine == 'x86_64'
types-python-dateutil==2.8.19.20240106 ; python_version >= '3.8'
typing-extensions==4.10.0 ; python_version >= '3.8'
tzdata==2024.1 ; python_version >= '2'
ultralytics==8.0.232
uri-template==1.3.0
urllib3==2.0.7 ; python_version >= '3.10'
uvicorn[standard]==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0 ; python_version >= '3.8'
websockets==12.0
widgetsnbextension==4.0.10 ; python_version >= '3.7'
